+++ src/main/java/com/yapp/web1/controller/testController.java
@@ -1,9 +1,11 @@
 package com.yapp.web1.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequestMapping("/api")
 public class testController {
 
     @GetMapping("/test")
